fix: bug sweep — display init, GPS baud timeout, DM name buffer, channel routing, save counter overflow (#632, #633, #634, #635, #637)
- #632: check tft.init() return value in display_init(), enable restart-on-fail path
- #633: add 60s GPS baud probe timeout to stop cycling when no GPS hardware connected
- #634: widen dyn_channels buffer [32]→[37] to prevent DM channel name truncation
- #635: reject empty/null channel names in find_channel_idx() instead of routing to active channel
- #637: uint8_t→uint16_t for save_counter to prevent 5-min persistence gap after ~2h uptime

All callers of exportChannels() updated to match new buffer width.
Tests: 774 test cases — 1 skipped, 773 passed.
Build: SigurdOS_TDeck SUCCESS, 2.45MB / 15.3% flash.

+++ src/ui/ui.cpp
@@ -141,7 +141,7 @@ void loop()
             home_screen_update_time(tbuf);
         }
         // Persist state every 5 min (catches unexpected power loss)
-        static uint8_t save_counter = 0;
+        static uint16_t save_counter = 0;
         if (++save_counter >= 10) {
             save_counter = 0;
             sigurdos::mesh::saveState();
